Day 12: Cleanup to separately solve parts 1 and 2
I realized the Visited map will have stored the distances from E to every interesting grid point
if I just let it search for S. This means it already has the distances for every 'a' cell, too.  So
to separately find the distance for S or a, we can just search for the shortest distance with 'S'
for part1 or the shortest distance for any 'a' or 'S' for part2.

It's a little extra work for part1 because the nav function directly returns the distance for 'S' anyway.
But if I fix that, I lose my generic search function at the end, which was a nice Rust trick.  :-(