Day16 revisited: measure paths independently
A new tactic runs 300x faster.  In this approach, we just collect all the combinations
of valves we can open in 26 minutes or less. Then we consider all the pairs of paths
we could take which do not open the same valve.  We add the combined pressure from
those pairs of paths and find the maximum.  And the number is right.  Yay!